+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Make the appearance of keys, prefixes, command descriptions and the divider configurable
+
+### Changed
+
+- By default display keys in bold and prefixes in the current theme's keyword color
+
 ## 0.1 - 2021-01-31
 
 ### Added